requirements.txt
Django>=5.2
djangorestframework>=3.14.0
django-extensions>=3.2.3
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
Pillow>=10.2.0